Fold ContractHelper.CoreCLR back into ContractHelper.
